refractor: setAlignHorizontalCenter move logic
