+++ frontend/src/pages/TemplateCreator/TemplateCreatorPage.tsx
@@ -30,7 +30,7 @@ import { delay } from "../../util/timeUtil";
 import { getPdfDimensions } from "../../util/fileGetUtil";
 import { getPlugins } from "../../util/pdfmeUtil";
 import { cloneDeep } from "../../util/fileActionUtil";
-import { setAlignVerticalCenter } from "./templateCreatorMutators";
+import { setAlignHorizontalCenter, setAlignVerticalCenter } from "./templateCreatorMutators";
 
 type Props = {
   templates: TemplateResponse[] | null;
@@ -553,21 +553,21 @@ useEffect(() => {
   };
 
   // Hardcoded width value of 215 for a4
-  const setAlignHorizontalCenter = () => {
-    if (designer.current && selectedField && pdfSize) {
-      // @ts-ignore
-      const selectedFieldWidth = designer.current.template.schemas[0][selectedField].width;
+  // const setAlignHorizontalCenter = () => {
+  //   if (designer.current && selectedField && pdfSize) {
+  //     // @ts-ignore
+  //     const selectedFieldWidth = designer.current.template.schemas[0][selectedField].width;
 
-      const centerPosition = ((calculateCanvasSizeFromPdfSize(pdfSize.width)) - selectedFieldWidth) / 2;
-      // @ts-ignore
-      designer.current.template.schemas[0][selectedField].position.x = centerPosition;
-      // @ts-ignore
-      designer.current.updateTemplate(designer.current.template);
+  //     const centerPosition = ((calculateCanvasSizeFromPdfSize(pdfSize.width)) - selectedFieldWidth) / 2;
+  //     // @ts-ignore
+  //     designer.current.template.schemas[0][selectedField].position.x = centerPosition;
+  //     // @ts-ignore
+  //     designer.current.updateTemplate(designer.current.template);
   
-      setTemplateStyle(prevState => ({ ...prevState, positionX: centerPosition }));
-      setFieldsChanged(true)
-    }
-  };
+  //     setTemplateStyle(prevState => ({ ...prevState, positionX: centerPosition }));
+  //     setFieldsChanged(true)
+  //   }
+  // };
   
   // Hardcoded width value of 305 for a4
   // const setAlignVerticalCenter = () => {
@@ -716,7 +716,7 @@ useEffect(() => {
                                 setPositionY={setPositionYHandler}
                                 setSizeWidth={setSizeWidthHandler}
                                 setSizeHeight={setSizeHeightHandler}
-                                setAlignHorizontalCenter={setAlignHorizontalCenter}
+                                setAlignHorizontalCenter={() => setAlignHorizontalCenter(designer, selectedField, pdfSize, setTemplateStyle, setFieldsChanged)}
                                 setAlignVerticalCenter={() => setAlignVerticalCenter(designer, selectedField, pdfSize, setTemplateStyle, setFieldsChanged)}
                                 fieldWidth={fieldWidth}
                                 fieldHeight={fieldHeight}
