+++ frontend/src/pages/TemplateCreator/templateCreatorMutators.ts
@@ -4,7 +4,8 @@ import { calculateCanvasSizeFromPdfSize } from "../../util/templateCreatorUtil";
 
 export const setAlignVerticalCenter = (
   designer: React.MutableRefObject<Designer>, 
-  selectedField: string, pdfSize: pdfSize, 
+  selectedField: string, 
+  pdfSize: pdfSize, 
   setTemplateStyle: (value: React.SetStateAction<TemplateInstanceStyle>) => void, 
   setFieldsChanged: (changed: boolean) => void
 ) => {
@@ -18,6 +19,29 @@ export const setAlignVerticalCenter = (
     designer.current.updateTemplate(designer.current.template);
 
     setTemplateStyle(prevState => ({ ...prevState, positionY: centerPosition }));
+    setFieldsChanged(true)
+  }
+};
+
+
+export const setAlignHorizontalCenter = (
+  designer: React.MutableRefObject<Designer>, 
+  selectedField: string,
+  pdfSize: pdfSize, 
+  setTemplateStyle: (value: React.SetStateAction<TemplateInstanceStyle>) => void, 
+  setFieldsChanged: (changed: boolean) => void,
+) => {
+  if (designer.current && selectedField && pdfSize) {
+    // @ts-ignore
+    const selectedFieldWidth = designer.current.template.schemas[0][selectedField].width;
+
+    const centerPosition = ((calculateCanvasSizeFromPdfSize(pdfSize.width)) - selectedFieldWidth) / 2;
+    // @ts-ignore
+    designer.current.template.schemas[0][selectedField].position.x = centerPosition;
+    // @ts-ignore
+    designer.current.updateTemplate(designer.current.template);
+
+    setTemplateStyle(prevState => ({ ...prevState, positionX: centerPosition }));
     setFieldsChanged(true)
   }
 };
